+++ src/com/rapidminer/example/table/LongSparseArrayDataRow.java
@@ -80,7 +80,7 @@ public class LongSparseArrayDataRow extends AbstractSparseArrayDataRow {
     /** Sets the given data for the given attribute. */
     @Override
 	public void setValue(int index, double v) {
-        values[index] = (short)v;
+        values[index] = (long) v;
     }
 
     @Override
